exception fixes

+++ src/main/java/fullstack/service/TalkService.java
@@ -80,21 +80,24 @@ public class TalkService implements PanacheRepository<Talk> {
     }
 
     @Transactional
-    public void deleteById(String sessionId,String id) throws SessionException {
+    public void deleteById(String sessionId,String id) throws SessionException, NoContentException {
         if (userService.isAdmin(sessionId)) {
             throw new AdminAccessException(ADMIN_REQUIRED);
+        }
+        if(talkRepository.findById(id) == null) {
+            throw new NoContentException(TALK_NOT_FOUND);
         }
         talkRepository.deleteById(id);
     }
 
     @Transactional
-    public void updateTalk(String sessionId, String id, Talk talk) throws SessionException {
+    public void updateTalk(String sessionId, String id, Talk talk) throws NoContentException {
         if (userService.isAdmin(sessionId)) {
             throw new AdminAccessException(ADMIN_REQUIRED);
         }
         int updated = update(id, talk);
         if (updated == 0) {
-            throw new SessionException(TALK_NOT_FOUND);
+            throw new NoContentException(TALK_NOT_FOUND);
         }
     }
 }
